+++ style.ts
@@ -5,7 +5,7 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     justifyContent: "flex-end",
-    backgroundColor: "#110436",
+    backgroundColor: "#E5C5BD",
   },
   containerTwo: {
     flex: 1,
@@ -16,18 +16,35 @@ const styles = StyleSheet.create({
     ...StyleSheet.absoluteFillObject,
     objectFit: "cover",
   },
-  button: {
-    backgroundColor: "rgba(123, 104, 238, 0.8)",
+  buttonWhite: {
+    backgroundColor: "#fff",
     height: 55,
     alignItems: "center",
     justifyContent: "center",
-    borderRadius: 35,
+    borderRadius: 15,
     marginHorizontal: 20,
     marginVertical: 10,
     borderWidth: 1,
     borderColor: "#fff",
   },
-  buttonText: {
+  buttonBlue: {
+    backgroundColor: "#325ECF",
+    height: 55,
+    alignItems: "center",
+    justifyContent: "center",
+    borderRadius: 15,
+    marginHorizontal: 20,
+    marginVertical: 10,
+    borderWidth: 1,
+    borderColor: "#325ECF",
+  },
+  buttonTextBlack: {
+    fontSize: 20,
+    fontWeight: "600",
+    color: "black",
+    letterSpacing: 0.5,
+  },
+  buttonTextWhite: {
     fontSize: 20,
     fontWeight: "600",
     color: "white",
@@ -44,19 +61,19 @@ const styles = StyleSheet.create({
     borderColor: "rgba(0, 0, 0, 0.2)",
     marginHorizontal: 20,
     marginVertical: 10,
-    borderRadius: 25,
+    borderRadius: 15,
     paddingLeft: 10,
   },
   formButton: {
-    backgroundColor: "rgba(123, 104, 238, 0.8)",
+    backgroundColor: "#d39e90",
     height: 55,
     alignItems: "center",
     justifyContent: "center",
-    borderRadius: 35,
+    borderRadius: 15,
     marginHorizontal: 20,
     marginVertical: 10,
     borderWidth: 1,
-    borderColor: "#fff",
+    borderColor: "#f2e2de",
     shadowColor: "#000",
     shadowOffset: {
       width: 0,
